requirements.txt
# Основной NLP стек (из INFRA-001)
natasha==1.6.0

navec>=0.10.0
slovnet>=0.6.0
ipymarkup>=0.9.0

# Разработка
jupyterlab==4.1.0
pytest==8.3.0
black==24.4.2
isort==5.13.2
mypy==1.10.0

# Инфраструктура (Modal, Nebius, HTTP)
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.0
modal>=0.60.0

# Логирование и утилиты
rich==13.9.3
typer==0.12.5

# Оценка бенчмарков
# conllu==4.4.1
# scikit-learn==1.5.1
# pandas==2.2.0
# pandas<1.6.0,>=1.0.0  # совместимо с DeepPavlov

# deeppavlov>=1.0.0

requests>=2.28.0
tqdm>=4.64.0
conllu>=4.5.2
pyyaml>=6.0

razdel>=0.5.0